CR-65 Moved validation tests to existing test case
